+++ Riot/Modules/Authentication/AuthenticationViewController.m
@@ -59,6 +59,7 @@
 @property (nonatomic, readonly) BOOL isIdentityServerConfigured;
 @property (nonatomic, strong) KeyVerificationCoordinatorBridgePresenter *keyVerificationCoordinatorBridgePresenter;
 @property (nonatomic, strong) SetPinCoordinatorBridgePresenter *setPinCoordinatorBridgePresenter;
+@property (nonatomic, strong) KeyboardAvoider *keyboardAvoider;
 
 @end
 
@@ -168,6 +169,8 @@
 
     [self userInterfaceThemeDidChange];
     [self updateUniversalLink];
+    
+    _keyboardAvoider = [[KeyboardAvoider alloc] initWithScrollViewContainerView:self.view scrollView:self.authenticationScrollView];
 }
 
 - (void)userInterfaceThemeDidChange
@@ -271,6 +274,8 @@
 
     // Screen tracking
     [[Analytics sharedInstance] trackScreen:@"Authentication"];
+    
+    [_keyboardAvoider startAvoiding];
 }
 
 - (void)viewDidAppear:(BOOL)animated
@@ -296,6 +301,13 @@
     }
 }
 
+- (void)viewDidDisappear:(BOOL)animated
+{
+    [_keyboardAvoider stopAvoiding];
+    
+    [super viewDidDisappear:animated];
+}
+
 - (void)destroy
 {
     [super destroy];
@@ -314,6 +326,7 @@
 
     autoDiscovery = nil;
     _keyVerificationCoordinatorBridgePresenter = nil;
+    _keyboardAvoider = nil;
 }
 
 - (BOOL)isIdentityServerConfigured
@@ -1021,6 +1034,10 @@
             {
                 constant += customServersContainerFrame.size.height;
             }
+            else
+            {
+                constant += self.customServersTickButton.frame.size.height;
+            }
         }
     }
 
@@ -1122,6 +1139,13 @@
             NSLog(@"[AuthenticationVC] adminContact(%@) cannot be opened", adminContact);
         }
     }];
+}
+
+#pragma mark - UITextFieldDelegate
+
+- (void)textFieldDidBeginEditing:(UITextField *)textField
+{
+    [self.authenticationScrollView vc_scrollTo:textField with:UIEdgeInsetsMake(-20, 0, -20, 0) animated:YES];
 }
 
 #pragma mark - KVO